requirements.txt
Django==1.6.11
django-custom-user==0.5
django-extensions==1.5.2
South==1.0.2
